requirements.txt
numpy
pandas
scipy
scikit-learn
matplotlib
pyyaml
tqdm
torch>=2.1
einops
